Move UriReplacement to kiota cli commons
