+++ src/Microsoft.Graph.Cli.Core/Microsoft.Graph.Cli.Core.csproj
@@ -44,7 +44,7 @@
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="7.0.0" />
     <PackageReference Include="Microsoft.Graph.Core" Version="3.1.0" />
-    <PackageReference Include="Microsoft.Kiota.Cli.Commons" Version="0.5.1-preview.1" />
+    <PackageReference Include="Microsoft.Kiota.Cli.Commons" Version="0.5.2-preview.1" />
     <PackageReference Include="Microsoft.Kiota.Http.HttpClientLibrary" Version="1.2.0" />
     <PackageReference Include="Spectre.Console" Version="0.47.0" />
     <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
